Another attempt at curling some dependencies.
This is a bad idea, I know, but I don't want a lot of cruft kicking
around here.

Turns out that the download of JsTestDriver was redirecting to an HTML
page and the download of coverage was using a auth token.  Added a
static copy of coverage and fixed the link to JsTestDriver.