After an extensive argument with PhantomJS, I realize that I had to give in.  I wanted a nice task that allowed me to bring up JsTestDriver and PhantomJS as child prcoesses, but spawning PhantomJS via node was causing it to launch two processes: one for node and one for the PhantomJS script.  I could kill one process, but not both.
Switching to execFile solved this problem.  I still can't seem to pipe the resulting processes output to stdout and I can't unref it like I could with spawn, but the end result is actually a reasonable experience for the developer.

I removed the -u and -d options for the Cakefile and added a -s option so that my pre-commit hook output is reasonable.